requirements.txt
# Temel Bağımlılıklar
aiogram>=3.20.0  # Telegram Bot API framework
aiohttp>=3.11.18  # Async HTTP client (aiogram için gerekli)
aiosignal>=1.3.2  # Async sinyaller için
frozenlist>=1.6.0  # Immutable listeler için
magic-filter>=1.0.12  # Aiogram filtreleri için
pydantic>=2.11.5  # Veri doğrulama (aiogram için gerekli)
pydantic-core>=2.33.2  # Pydantic core
requests>=2.32.3  # Senkron HTTP istekleri için
filelock>=3.18.0  # Dosya kilitleme
python-dotenv>=1.1.0  # Çevresel değişkenler
httpx>=0.28.1  # Modern HTTP client
anyio>=4.9.0  # Async I/O desteği
openai>=1.14.0  # OpenAI API client

# Test Araçları
pytest>=8.1.1  # Test framework

# Loglama ve Metrik Sistemi
python-json-logger>=3.3.0
structlog>=25.3.0  # Yapılandırılmış loglama
attrs>=25.3.0  # Veri sınıfları için
propcache>=0.3.1  # Property caching

# Güvenlik ve Kimlik Doğrulama
pyjwt>=2.10.1
cryptography>=45.0.2
pyaes>=1.6.1  # AES şifreleme
rsa>=4.9.1  # RSA şifreleme
pyasn1>=0.6.1  # ASN.1 desteği

# Veritabanı ve Cache
redis>=6.1.0  # Redis client
python-redis-lock>=4.0.0  # Redis tabanlı distributed lock
portalocker>=3.1.1  # Cross-platform dosya kilitleme

# Dosya İşlemleri ve Performans
ijson>=3.4.0  # Büyük JSON dosyaları için streaming parser
zstandard>=0.23.0  # Yüksek performanslı sıkıştırma
jsonschema>=4.23.0  # JSON şema doğrulama
referencing>=0.36.2  # JSON şema referansları
rpds-py>=0.25.1  # Persistent veri yapıları
pathlib>=1.0.1  # Path işlemleri
typing-extensions>=4.13.2  # Gelişmiş tip tanımlamaları
aiofiles>=24.1.0  # Async dosya işlemleri

# Tip Kontrolü ve Geliştirici Araçları
mypy>=1.15.0
mypy-extensions>=1.1.0
types-requests>=2.32.0.20250515
types-redis>=4.6.0.20241004
types-jsonschema>=4.23.0.20250516
types-filelock>=3.2.7
types-setuptools>=80.8.0.20250521
types-cffi>=1.17.0.20250523
types-pyOpenSSL>=24.1.0.20240722
annotated-types>=0.7.0  # Type annotation desteği

# Not: asyncio, threading, multiprocessing, json, shutil, tempfile, 
# logging, time, datetime, os, typing gibi modüller Python'un 
# standart kütüphanesinin parçasıdır.

# Multi-Database Dependencies
sqlalchemy>=2.0.0  # ORM ve async database toolkit
asyncpg>=0.29.0  # Async PostgreSQL driver
motor>=3.6.0  # Async MongoDB driver
redis[hiredis]>=6.1.0  # Redis with hiredis parser
pymongo>=4.10.0  # MongoDB driver (motor dependency)

# Opsiyonel Bağımlılıklar (ihtiyaca göre yüklenebilir)
# psycopg2-binary>=2.9.9  # PostgreSQL bağlantısı için
# fastapi>=0.109.0  # API geliştirme için
# uvicorn>=0.27.0  # ASGI server 